% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv data from file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string containing Path to the file if file is not stored in current directory.
filename if the csv file is stored in current directory.}
}
\value{
This function returns a data frame containing data from the file read.
}
\description{
This function reads data from a csv file or a zipped file with csv data,if it exists.
If the file doesnt exist it gives an error message
saying the file \code{file filename does not exist} and stops the execution.
}
\details{
The function is a wrapper for read_csv from the readr package. The function throws an error if the file
with given filename is not found and stops program execution.
}
\examples{
fars_read('myfile.csv.bz')
fars_read('myfile.csv')

}
