% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Make a list of dataframe containing month and year data for accidents.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a  character or integer vector contaning years}
}
\value{
The function returns a list of dataframes one for each year of data considered .Each dataframe has two columns
corresponding to the month and year for each accident for that year.
}
\description{
This function reads a vector (\code{character,integer}) of years and returns a list of dataframes containing the
month of accident for each accident in the given year.If for any reason the program is unable to read
data for a paricular year an warning (\code{invalid year: year}) is displayed and a NULL is returned for that year's value
}
\details{
The function takes in a vector of years.It first uses the make_filename function to create a filename .The looks for a file with said filename
and reads data from it using fars_read, to retain only the month and year column before moving onto next year.
As described if the program is unable to read data for particular year it throws an warning message and returns a NULL for that year
}
