% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Make a data.frame containing number of accidents in a given month for each year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a  character or integer vector contaning years}
}
\value{
The function returns a dataframe with columns as years and rows as number of accidents during each month for that year.
}
\description{
This function reads a vector (\code{character,integer}) of years and returns a dataframe summarizing total number of accidents
for each month for given year in set of years. A warning message will be thrown if the program cannot
find data file for a particular year.
}
\details{
The function takes in a vector of years.Uses fars_read_year to create a list of dataframes containing accident month
and year data.Groups data by year and month and then summarizes it using dplyr functions and then creates dataframe with months as rows and a column
for each year using spread function.A warning message will be thrown if the program cannot
find data file for a particular year (\code{invalid year: year}). This behavior is derived from the fars_read_years function
}
