% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make csv.zip type filename from a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a integer,character value for the year under consideration}
}
\value{
The function returns a string vector containing the filename.The filename is of form (\code{accident_year.csv.bz2})
}
\description{
This function reads a year (\code{character,integer}) and creates a filename for a csv type zipped file containing the accident  data
for the year.
}
\details{
The function offers a convinient way to create file name to search for revelant data files
}
\examples{
make_filename(2013)
make_filename('2013')

}
